+++ CHANGELOG.md
@@ -12,6 +12,13 @@
 - Save real position in NBT instead of teleporting it?
 
 ----
+### 3.1.11
+##### Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2, 1.19, 1.19.1-1.19.2, 1.19.3, 1.19.4, 1.20[.1], 1.20.2, 1.20.3-1.20.4, 1.20.5-1.20.6, 1.21-1.21.1, 1.21.2-1.21.4, 1.21.5
+#### Fixes
+- Fix server-side translation, fix [#163](https://github.com/NikitaCartes/EasyAuth/issues/163)
+- Fix vehicle disappear if player log out in different dimension with hide-player-coords, fix [#141](https://github.com/NikitaCartes/EasyAuth/issues/141)
+
+----
 ### 3.1.10
 ##### Minecraft 1.17, 1.17.1, 1.18[.1], 1.18.2, 1.19, 1.19.1-1.19.2, 1.19.3, 1.19.4, 1.20[.1], 1.20.2, 1.20.3-1.20.4, 1.20.5-1.20.6, 1.21-1.21.1, 1.21.2-1.21.4, 1.21.5
 #### Fixes